Create Employee_gross wage
Task: Write an algorithm that compares the gross salary of each employee with the average gross salary and if the employee's salary is lower than the average gross salary, it adds €50 to the gross salary and displays the current adjusted salary. If the salary is equal to the average salary, it adds €20 to the employee and displays the adjusted salary, and if the employee has a salary higher than the average gross salary, the salary does not change and the algorithm displays it.